demos/a2a/requirements.txt
# Core dependencies extracted from pyproject.toml
# This file is generated for Docker builds that prefer pip over Poetry

# Main dependencies
google-adk>=1.0.0
google-genai>=1.9.0
pydantic>=2.10.6
python-dotenv>=1.0.1
google-cloud-aiplatform[adk,agent-engines]>=1.93
litellm>=0.10.0

# A2A Python SDK for agent-to-agent communication
a2a-sdk

# Web server dependencies for A2A
uvicorn>=0.24.0
fastapi>=0.104.0
starlette>=0.27.0
click>=8.0.0

# Additional dependencies that might be needed
requests>=2.25.0
typing-extensions>=4.0.0
